Basic start of the tool.  The focus is on audio, not graphics, so I don't mind the command-line here rather than all the effort of a GUI.

+++ Source/AudioSink.cpp
@@ -213,7 +213,7 @@ void AudioSink::GenerateAudio(double desiredSecondsAvailable, double minSecondsA
 		{
 			if (this->ownsStreamMemory)
 				delete audioStreamIn;
-			uint32_t j = this->audioStreamInArray->size() - 1;
+			uint32_t j = uint32_t(this->audioStreamInArray->size()) - 1;
 			if (i != j)
 				(*this->audioStreamInArray)[i] = (*this->audioStreamInArray)[j];
 			this->audioStreamInArray->pop_back();
